+++ src/interpolation.lua
@@ -3,6 +3,12 @@
 --- -----------------------------------------------------------
 
 
+--- @meta
+
+
+--- -----------------------------------------------------------
+
+
 --- contain a list of interpolation functions,
 --- each method require a start value, end value
 --- and the position in time from start to end of the interpolation (0.0 to 1.0).
@@ -10,7 +16,7 @@
 --- 	interpolation.linear(0, 8, 0.5) -- 4
 --- 	interpolation.quadratic_in(0, 8, 0.5) -- 2
 --- 	interpolation.quadratic_out(0, 8, 0.5) -- 4.5
---- @type table<function>
+--- @type { [string]: fun(x: number, y: number, t: number) }
 BeTweenApi.interpolation = {}